Fix part of the issue of the 'complex case not timing out' test which has apparently been broken for a while.
This is part of the fix. DuplicationPower wasn't in place yet when test was created, and it's reasonable to open with Bash+ if you've got duplication power cause I think you can kill then.
This fixes the test for a while after the commit that added DuplicationPower. But there's something still broken about it elsewhere. Can't narrow it down after quite some time fiddling.